add initial files for taskPicker activity using mvp architecture. set manifest to start this activity
