+++ app/src/main/java/com/jbnm/homehero/ui/taskpicker/TaskPickerMvpView.kt
@@ -1,0 +1,7 @@
+package com.jbnm.homehero.ui.taskpicker
+
+import com.jbnm.homehero.data.model.Task
+
+interface TaskPickerMvpView {
+    fun addTasks(tasks: List<Task>)
+}
