+++ app/src/main/java/com/jbnm/homehero/ui/taskpicker/TaskPickerPresenter.kt
@@ -1,0 +1,26 @@
+package com.jbnm.homehero.ui.taskpicker
+
+import com.jbnm.homehero.data.model.Task
+
+class TaskPickerPresenter(var mvpView: TaskPickerMvpView) {
+
+
+    fun loadTasks() {
+        mvpView.addTasks(generateTasks())
+    }
+
+    fun generateTasks(): List<Task> {
+        val instructions: List<String> = listOf("step1", "step2")
+        val tasks: List<Task> = listOf(
+                Task("testId", "Task 1", instructions, true),
+                Task("testId", "Task 2", instructions, false),
+                Task("testId", "Task 3", instructions, true),
+                Task("testId", "Task 4", instructions, true),
+                Task("testId", "Task 5", instructions, true),
+                Task("testId", "Really long chore description", instructions, true),
+                Task("testId", "Task 7", instructions, false)
+        )
+        return tasks
+    }
+
+}
