+++ app/src/main/res/layout/activity_task_picker.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.jbnm.homehero.ui.taskpicker.TaskPickerActivity">
+
+    <com.jbnm.homehero.ui.taskselector.TaskSelector
+        android:id="@+id/taskSelector"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:padding="30dp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:taskTextColor="@color/colorPrimaryDark"
+        app:taskTextSize="18sp"
+        app:colorOne="@color/taskWheelColorOne"
+        app:colorTwo="@color/taskWheelColorTwo"/>
+
+</android.support.constraint.ConstraintLayout>
